web/requirements.txt
alembic==1.11.3
APScheduler==3.10.3
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Flask==2.3.2
Flask-APScheduler==1.12.4
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
gtfs-realtime-bindings==1.0.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
pip-review==1.3.0
postgres==4.0
protobuf==4.24.1
psycopg2==2.9.7
psycopg2-binary==2.9.7
psycopg2-pool==1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
Werkzeug==2.3.7
zipp==3.16.2
